// src/worldcomm_pb.d.ts
export enum FlowStatus {
  UNKNOWN_STATUS = 0,
  OPEN = 1,
  CLOSE = 2,
}